Move all dsl fixtures into conftest
